* move phpunit test files to match with current namespace hierarchy * rename phpunit test classes to match with current systems under test * replace all references to `Tests\TestCase` in `be` with `Symfony\Bundle\FrameworkBundle\Test\KernelTestCase` @ `App\Tests` @ symfony
