+++ symfony/tests/PostsQuery/QueryParamsTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Tests\Http\PostsQuery;
+namespace App\Tests\PostsQuery;
 
 use App\Helper;
-use App\Http\PostsQuery\QueryParam;
-use App\Http\PostsQuery\QueryParams;
+use App\PostsQuery\QueryParam;
+use App\PostsQuery\QueryParams;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
